refactor(api): move TokenBufferMemory to model_runtime
